a ruby wrapper that launches ruby_drills
I barely have an idea of how this code works. I cargo-culted it from this
stack overflow post: http://stackoverflow.com/questions/2730642/wrapper-around-bash-control-stdin-and-stdout

Here is how I think it works.
In any command line program, I type stuff in, and I get stuff out.
when I type stuff in, it's like I'm dropping rock into a pipe bucket called standard-in
when I press enter, all of the rocks get dumped into a mysterios machine named ruby_drills
he machine grumbles, and belches out a bunch of rocks called stdout.
reading what the program gives me is like pocking through the rocks in the stdout bucket.

If ruby drills us a machine, I am the poor factiry worker who has to deal with tge things it is putting in
and getting out.  Now, I get bored shoveling rock in and getting it out all day, so I am going to make a computer do it for me.
all the boring work.

So this program is a robot that shovels the rock in for me and gives them back out.

Right now, I have a robot.  I don't know how it works, cuz I got it fir free from a random perdon on the internet. I am aldo handing it rocks and taking rocks back from it. I even put a box over the robot so that when I'm handing it things it still looks like the old robot. So everything looks the sane, I just have a secret robot doing things.